+++ CHANGELOG.md
@@ -10,8 +10,9 @@ All notable changes to this project will be documented in this file.
 
 ### Deprecated
 
-* Deprecated `AbstractModelCollection->set_item()` because of its misleading naming. Use `AbstractModelCollection->add()` instead.
-* Deprecated `AbstractModelCollection->remove_item()` because of its inconsistent naming. Use `AbstractModelCollection->remove()` instead.
+* Deprecated `AbstractModelCollection->get_by_id()`. Use `AbstractModelCollection->getById()` instead.
+* Deprecated `AbstractModelCollection->remove_item()`. Use `AbstractModelCollection->remove()` instead.
+* Deprecated `AbstractModelCollection->set_item()`. Use `AbstractModelCollection->add()` instead.
 * Deprecated `BaseHttpAuthentication->check_authentication()`. Use `BaseHttpAuthentication->checkAuthentication()` instead.
 * Deprecated `BaseSessionInterface->clear_all()`. Use `BaseSessionInterface->clearAll()` instead.
 * Deprecated `Configuration::remove_key()`. Use `Configuration::remove()` instead.